refactor(frontend): rename file `useManifest.ts` -> `manifest.ts`
