+++ frontend/src/queries/manifest.ts
